move deprecated level loader into services
